+++ playground/grgraphics/simple_backscroll.cpp
@@ -178,7 +178,7 @@ public:
             m_lastAbsVelocity = absVelocity;
             m_xVelocity *= speedRamp;
             //m_csvFile.WriteData(m_xVelocity);
-            grcore::writeData(m_xVelocity);
+            grcore::writeData(static_cast<const double>(m_xVelocity));
 
             { // handle Graph
                 m_view.move(m_xVelocity * elapsedSeconds, 0.0F);
